Avoid downcast

+++ core/client/client.go
@@ -157,7 +157,7 @@ func (c *Client) GetEntry(ctx context.Context, userID, appID string, opts ...grp
 	return e.GetCommitted().GetData(), slr, err
 }
 
-func min(x, y int32) int32 {
+func min(x, y int64) int64 {
 	if x < y {
 		return x
 	}
@@ -175,8 +175,8 @@ func (c *Client) PaginateHistory(ctx context.Context, appID, userID string, star
 	epochsWant := end - start + 1
 	for int64(len(allProfiles)) < epochsWant {
 		log.Printf("pageSize: %v", c.pageSize)
-		count := min(int32(epochsWant-int64(len(allProfiles))), c.pageSize)
-		profiles, next, err := c.VerifiedListHistory(ctx, appID, userID, start, count)
+		count := min(epochsWant-int64(len(allProfiles)), int64(c.pageSize))
+		profiles, next, err := c.VerifiedListHistory(ctx, appID, userID, start, int32(count))
 		if err != nil {
 			return nil, nil, fmt.Errorf("VerifiedListHistory(%v, %v): %v", start, count, err)
 		}
